Modify IAudioPlayerNode to start in main thread
The AudioJitterBuffer should reset when starting by getting the SSRC
refreshing frames from the previous frame, but it can be abandonded when
the IAudioPlayerNode started with delayed.
I fix the problem to make the IAudioPlayerNode start from the main
thread sync with other nodes to start together to make the frame
delivering started after the IAudioPlayerNode finised to start.

Bug: 289874039
Test: Verified voice call in live network and TE test cases, atest ImsMediaNativeTests
Change-Id: Ia0fe47c59220273855139d8d16667f083bc297f3