Modify to send audio frame type from the IAudioPlayerNode
The numNoDataFrames are always reported 0 in the CallQuality report. it has to be more than 0 when there is gap between voice frame every 20 msec. To report the non zero NoDataFrames, it has to be triggered in IAudioPlayerNode when it gets the audio frames from the AudioJitterBuffer and report the NoDataFrames when the AudioJitterBuffer returns false when there is no frame to play.

Bug: 291103302
Test: Verified by the voice call in live network, atest ImsMediaNativeTests
Change-Id: If3fa32bb8064d3572d8b983b2ffef9b03371bf2c