Fix the uninteded handover by the incorrect loss rate calculation.
When there is a event of the deleting late arrival of the DTX frame, and
there is a voice frame played for long time until there is no frame to
get, the preserved frame used and the lost frame switched by its frame.
It causes the incorrect lost frame counter to make this handover. To fix
this, add logic to delete the preserved dtx frame to use later when the
Get frame is called with valid frame.

Bug: 301104195
Test: Verified by the VoWifi call in live network
Change-Id: I67c3be01ab27cc678cb20dd6f3ec63fe9013531d